revised app.js
