+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "moment": "^2.20.1",
     "react": "15.6.1",
     "react-bootstrap": "^0.32.1",
     "react-dom": "15.6.1",
